+++ tests/test_consumer.py
@@ -9,6 +9,7 @@ from unittest.mock import Mock, call
 
 import pytest
 from streaming_kafka_consumer import Message, Partition, Topic
+from streaming_kafka_consumer.strategy_factory import KafkaConsumerStrategyFactory
 
 from snuba.clusters.cluster import ClickhouseClientSettings
 from snuba.consumers.consumer import (
@@ -23,7 +24,6 @@ from snuba.datasets.storage import Storage
 from snuba.processor import InsertBatch, ReplacementBatch
 from snuba.utils.metrics.wrapper import MetricsWrapper
 from snuba.utils.streams.backends.kafka import KafkaPayload
-from snuba.utils.streams.strategy_factory import KafkaConsumerStrategyFactory
 from tests.assertions import assert_changes
 from tests.backends.confluent_kafka import FakeConfluentKafkaProducer
 from tests.backends.metrics import TestingMetricsBackend, Timing
